Menu Instructions / minor renaming
