+++ CSE442_Project/Assets/Fonts.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: d3675e3dd6cb942e08dcfa01400502a1
+folderAsset: yes
+timeCreated: 1519087752
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
