+++ CSE442_Project/Assets/_Scenes/MENU-Instructions.unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 8283c066d15154124b7e3d26e191dca7
+timeCreated: 1519349386
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
